Making cc/tiles use SkColor4f
This CL is part of the effort of increasing color precision and
color gamut support in chrome.
https://docs.google.com/document/d/1y5qhBuK7sFkXt7c6vcrAJeIofj3h2UQ3btX3Hv1hrz8/edit#heading=h.69111d60ji6h
particularly
https://docs.google.com/document/d/1kvl_hwnIG3lAXcSxRndgcqbDGEvGoiD52-VwOuVfveE/edit?resourcekey=0-uvlZUrkCQ-o0H8ohYepsqA
To do so we are changing viz, cc, and other chrome parts to use
SkColor4f instead of SkColor.

This CL finalizes the last code in cc/tiles that was using SkColor and
changes it to SkColor4f.

With this, cc/tiles should be SkColor4fized.

Bug: 1308932
Change-Id: I82bb480c097d913bf66506037e8992bd5a6c1742
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3769274
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025320}