Intents: Flip intent chip logic to be based around 'expanded' state
Currently, the intent chip logic considers whether the chip is
'collapsed' and should be shown with an icon instead of an entire chip.
This CL flips the logic, so that the code considers whether the chip
should be shown as 'expanded'. This is mostly just a refactoring/code
clarity change.

There is one behavior impact, which is that the default value now
considers the chip to be collapsed (that is, the 'expanded' state
defaults to false) rather than expanded. This only impacts the ARC
external protocol dialog when the LinkCapturingUiUpdate feature is
enabled, and it changes the behavior to match what happens when the
feature is disabled.

Bug: 1300776
Change-Id: I927182a85413eb21cb7d59f64f4ef0326d6c9a55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3759050
Auto-Submit: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025333}