[CacheStorage] 2/ Add StorageKey unit tests
This CL is the second in a set of CLs to integrate Cache Storage with the
Buckets API. The buckets integration is nearly complete in a WIP CL [1],
but we are splitting up the changes into smaller CLs for easier review.

This CL updates some existing tests and adds new tests specifically
around the existing StorageKey partitioning functionality.

Design doc:
https://docs.google.com/document/d/1VDm8M4vRyxCpvp_m3Ojs18W9g1WpqvSOsKf0A6hnAFM/edit?pli=1#heading=h.1kp8e5sw3qha

Implementation design doc:
https://docs.google.com/document/d/1B_tUfTJDeuqRyehvX6aM44YwvOkKWVNM5LNBu3_leqU/edit?usp=sharing

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3735215

Bug: 1218097
Change-Id: I35cdcc9af6c98385d6b6ae3cab22671616f6cea0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3764140
Commit-Queue: Andrew Williams <awillia@google.com>
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025379}