[side search] Refactor Side Search tests p4
This duplicates relevant tests from the SideSearchBrowserControllerTest
to the SideSearchV2Test test suite.

It may be possible to further parametrize the tests in the browser
controller test suite - however this adds additional complexity,
makes the tests harder to read and there may be slight behavioral
differences between the implementations. Further, not all tests
(focus tests for e.g.) should be run for both implementations.

The following tests were excluded when porting tests to the unified
test suite:

  - TimeUntilOpenMetricEmittedCorrectlyMultipleNavigations
  - SwitchingTabsHandlesFocusCorrectly
  - OpeningAndClosingTheSidePanelHandlesFocusCorrectly
  - ReadLaterWorkInOTR

This CL also addressed a bug where the UnifiedSideSearchController
was not setting itself correctly as a Delegate to the
SideSearchTabContentsHelper. This was because calls to
WebContentsUserData::FromWebContents() in the constructor would
always return null (due to the fact the user data has not been set
at that point).

The test that handles crash handling has been disabled. Follow
up CLs should investigate what the expected behavior for
side panel WebContents crashes in the foreground and
background should be for the unified side panel and re-enable
this test.

Bug: 1343667
Change-Id: Ief31b06291e136dae7964250a4631245366e4fcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3763211
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025380}