[ZPS] Adds a new ZPS prefetching escape hatch feature for Android
If both the prefetching feature and this new default-enabled feature
are enabled, Android calls AutocompleteController::StartPrefetch()
on the same instance as the one used by the omnibox.

If only the prefetching feature is enabled, Android uses
ZeroSuggestPrefetcher which spins off a new AutcompleteController instance; however calls StartPrefetch() to be more in line with the
other platforms.

Otherwise, Android uses ZeroSuggestPrefetcher and calls Start() which
is the default behavior.

ZeroSuggestPrefetcher is largely simply refactored less two changes:
1) The generous 10 sec self-destruction timeout is changed to the value
   used by the AutocompleteController to stop the controller. This is
   sufficient since the provider will be stopped afterward anyway.
2) Uses a weak_ptr to bind the timer callback for extra safety.

Bug: 1262373
Change-Id: If4a3315769b6678a8d57879500aeb92c0337e019
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3758481
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025389}