views: fix updating selection state when adding ImageView
MenuItemView caches some state based on the selection. It does
this by checking if the selection changed at various point and
if so updating the state.

One of the things the selection is based off is
NonIconChildViewsCount(), which checks the number of views. To
handles changes to the number of child views MenuItemView
overrides ViewHierarchyChanged() and calls to
UpdateSelectionBasedStateIfChanged(). 
The implementation of NonIconChildViewsCount() checks some
member fields. These fields are typically
assigned using the pattern:

  foo_view_ = AddChildView(...);

The problem is ViewHierarchyChanged() is called before the
member has been assigned, so that NonIconChildViewsCount()
may return the wrong thing. To account for this, this patch
adds a field that makes ViewHierarchyChanged() do nothing
when set, so that the calling code can explicitly call
UpdateSelectionBasedStateIfChanged() after the member has been
assigned.

This only applies to fields that are set outside of the constructor
as UpdateSelectionBasedStateIfChanged() does nothing if not
contained within a widget.

BUG=1342390
TEST=covered by test

Change-Id: I8d43f05dd97b2647b581d69ca3913784f5b63a4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3763707
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025393}