[side search] Update side panel v2 text with hosted content domain
This CL updates side search combobox and header text to reflect the
domain of the content hosted in the side search side panel.

This text is the same that is already used for the existing side
search side panel.

Google DSE:
https://screenshot.googleplex.com/35QzYVYYC6cLpGe.png

Third Party DSE:
https://screenshot.googleplex.com/BwwnpXhu5Qh4q5L.png

Bug: 1343667
Change-Id: Idaa31a123ba2918a7f99da971b9ee057fc6a109a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3761210
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025446}