Store sticky constraints in (LayoutObject's) FragmentData::RareData
This avoids the need of PaintLayer::AncestorScrollContainerLayer()
when we access sticky constraints. PaintLayerScrollableArea now stores
a set of contained sticky layers instead of a map from sticky layers
to sticky constraints.

Add StickyPositionScrollingConstraints::constraining_rect (result of
LayoutBox::ComputeStickyConstrainingRect -- previous
LayoutBoxModelObject::ComputeStickyConstrianingRect) so that when it's
needed, we don't need to compute it again with
AncestorScrollContainerLayer().

This prepares for the change of letting AncestorScrollContainerLayer()
follow the containing block chain, which would make it difficult to
cache AncestorScrollContainerLayer(), especially to invalidate the
cache when PaintLayer tree structure changes. Based on this CL,
AncestorScrollContainerLayer() will always look up the tree without
caching, with a bigger cost of itself (but smaller cost to invalidate
and refresh the cache). Reducing the need of
AncestorScrollContainerLayer() will mitigate the performance impact.

Bug: 881555
Change-Id: I7b78b3ce05bd9578a495cda137fab563b128893b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3759984
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025452}