Reland "Exo: Rename exo::Capabilities -> exo::SecurityDelegate"
This is a reland of commit 727470b74681448c8a9e650103fb682f7705c831

The issue was a memory leak of the wayland client objects. Previously the CapabilityBindingTest would run first, and the leaks would be hidden by a crash in the following test. The cl was only a rename, but the effect is that SecurityDelegateBindingTest runs last, meaning that the crash no longer hides the memory leak.

The leak was in unit-test code only and is not a production concern. The fix was to clean up the wayland client objects correctly.

Original change's description:
> Exo: Rename exo::Capabilities -> exo::SecurityDelegate
>
> This is purely a rename+comment update. No functional changes.
>
> Formalising go/secure-exo-ids as "capabilities" makes the implementation
> difficult, since capabilities are (1) fine-grained, (2) dynamic and (3)
> difficult to associate with wayland objects (like shell surfaces).
>
> Instead, exo will use a "security delegate", which greatly simplifies
> exo's implementation.
>
> Bug: b/190141156
> Change-Id: Ibe5b4b29e3b5f52ef8f5fd174da6f4e2fb8cd4da
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3757039
> Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
> Commit-Queue: Nic Hollingum <hollingum@google.com>
> Reviewed-by: David Munro <davidmunro@google.com>
> Cr-Commit-Position: refs/heads/main@{#1024055}

Bug: b/190141156
Change-Id: I2ad512d3beba2e47961e362b8c8e2924c801bffd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3764780
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025470}