[Read Anything] Remove duplicate calls to DistillAXTree.
Before this CL, we called ReadAnythingController::DistillAXTree when:
 - The UI is ready
 - The UI is activated
 - The active tab changed
 - The active web contents finished loading

Sometimes multiple of these happened at the same time so we made
multiple consecutive calls to DistillAXTree which is wasteful. This
CL ensures that we only distill once for the currently active web
contents, and that we do not distill until the UI is ready.

Bug: 1266555
Change-Id: I10f59be26c2d88bdc8d81d3d14fcc011e765526a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3768648
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1025489}