Intents: Always show intent chip as expanded if the app is preferred
When the Intent Chip is shown in the omnibox for a single app, if that
app has the supported links preference enabled through Settings, always
show the chip in expanded style. Enabling the link capturing preference
for that app is a strong signal that the app is useful compared to a
browser tab.

Bug: 1300776
Change-Id: Id42e3ec8f2e6447f4746c6bd245e1cd16b719536
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3760254
Auto-Submit: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025491}