Reinstate "Delete" command.
The simple "delete" command is our second most used command. It will
be easier for users to continue to say "delete" than to say "delete
the previous character". That longer phrase could be supported
separately if we want to do so with regexes, and will also come
along with Pumpkin.

AX-Relnotes: Fixes bug in current branch where "delete" was not recognized.
Test: Dictation browsertest
Bug: 1344883
Change-Id: I2af53728c02ec9157c60641e35c9d46e68ee13d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3764896
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025519}