[Read Anything] Only distill AXTree when UI exists.
Before this CL, we distilled the AXTree when it was created. This CL
ensures that we distill the AXTree every time that the UI is created,
and prevent AXTrees from being distilled when the UI is destroyed.

The difference between the UI being activated vs. ready:
 - Activated: UI is visible in the Side Panel
 - Created: UI classes are constructed

Bug: 1266555
Change-Id: I5231e1b83da28d8f6289504a7570bcf1afbc4c28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3770174
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1025524}