Don't reset CompositorSurfaceManager if surface has been created
CompositorView has maybeResetCompositorSurfaceManager workaround that
reset surfaces when screen is off and on again.

Reset is implemented to run on surfaceDestroyed(), but usually it's
executed on surfaceCreate() like these.
1. surfaceDestroyed comes eariler than screen off msg
2. screen off msg comes next, so mNeedsReset is cached
3. When CompositorView want to request another surface
   (e.g., text selection)
3. another surface's surfaceCreated() happen and it trigger
   CompositorView's surfaceDestroyed() for previous surface.
4. surfaceDestroyed() reset surfaces because mNeedsReset was cached.

This flow can make the following problem cases.

Case 1. Screen blink when creating new surface
When resetting surfaces, old surface is detached and new surface is
attached. During detaching surface, framework  trigger one draw, so
the draw show blank frame to the screen, so user can see blinking.
Framework's RenderNode's onPositionLost() trigger the draw.

Case 2. surfaceRedrawNeededAsync mismatch and make the view stuck
If a surface is reset during creating the surface, mPendingReportDraw
of SurfaceView will not be 0, so the view will be stuck. When surface
is detaching from window, the count become 0, but our code is still in
surfaceCreated() callback, so if we return the callback successfully,
SurfaceView increase the count right after the callback. The count
never decrease to 0 because the view is already destroyed.
Now CompositorSurfaceManagerImpl force to make exception to avoid
the situation.

This patch don't reset the surfaces if surface is created successfully
when screen is on again.

Bug: 1344407
Change-Id: Ib538f7b66d513323547b345076ad37d395dd6484
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3760059
Commit-Queue: Sangheon Kim <sangheon77.kim@samsung.com>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025534}