Disable commands when Dictation language doesn't match system language.
In the case where the user's Dictation language doesn't match the
system language we have been showing commands and looking to recognize
commands in the system language even though speech recognition is
occurring in a different language. This means that commands are not
recognized and hints are shown in the wrong language, leading to a
bad user experience.

Until we can get Dictation command strings in the Dictation language,
disable commands when the two languages don't match.

This change compares the language part of the Dictation locale to
the language part of the system language to determine if we are
in a supported state. This way 'es-ES' matches both 'es-PR' and 'es-US'
for example. A few Dictation locales actually map to system locales
even though the language part of the locale does not match, and these
are special-cased.

Googlers, see go/dictation-vs-system-language-bug for impact and more.

AX-Relnotes: Removes commands UI from Dictation when commands are not in Dictation language.
Bug: 1340590
Change-Id: I89f8b66059089f545704c9cc84acaabfbc88b6ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3770275
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025549}