prerender: Reland "Make WebNavigation be aware of prerendering."
This is a reland of commit I5d46eead97eafce90dd4988b6200a55bf026fe4d

There are two major points for relanding.
1. This CL use executeScript for activation as chrome.tabs.update doesn't trigger activation. And crrev.com/c/3744705 allow prerendering
activation to be triggered by extensions.
2. There was a crash reported in script_executor, but the changes are
handled in crrev.com/c/3755182. So there is no related change in this
CL for the part.

"This CL makes WebNavigation be aware of prerendering frames and add a
relate test."

Bug: 1278141
Change-Id: I1da3ff50326e18334682353bc333de3e31eb3ff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3750758
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Huanpo Lin <robertlin@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025553}