dPWAs: Improve parameter naming in ExternallyManagedAppManager
This CL updates ExternallyManagedAppManager code to call "app_url"
"install_url" instead to be more specific and match the name in
ExternalInstallOptions::install_url.

This also adds names to the key/value parameters of the synchronise
install and uninstall results to make it clearer that the keys are the
install URLs.

Change-Id: Iad4b06e1026692a81b4470ea0724f45853c205a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3767772
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Auto-Submit: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Glen Robertson <glenrob@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025622}