Clean stale user for ephemeral mount
Make the Remove request to cryptohome in case during an ephemeral user
login an existing user homedir is discovered.

This fixes inability to log in in case there's a stale persistent user
that cryptohome hasn't had a chance to remove (e.g., because the policy
was updated while the device was at the Login Screen).

Bug: 1345476
Bug: b/235811135
Test: enroll DUT, login, logout, enable ephemeral policy, wait a few minutes on the login screen, click "add user" and enter the same user's credentials, verify logged in successfully

Change-Id: I2f96d92651e522c93cf2c7123a2f283f0bca763e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3771342
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025693}