Fenced frame: Add API to tell if opaque ad can load
This CL introduces a new static API call for HTMLFencedFrameElement
to determine if an opaque-ads fenced frame is allowed to load in
a specified context. The API will look at the current execution context
and determine (based on CSP headers, CSPEE set in ancestor iframes, fenced frame tree mode, secure
context, and sandbox flags) if an opaque-ads fenced frame would be
allowed if a script attempted to create one.

On the browser side, since the renderer doesn't have access to all
the information it needs, this CL adds browser-side calculations
(currently just CSPEE, but with the idea that it can be expanded in the
future) to tell frames if they can load a fenced frame.

Note that no fenced frames need to be created to invoke this method. It
can be called from the web platform with
HTMLFencedFrameElement.canLoadOpaqueURL();

See comment thread discussing the need for this API in the first place:
https://docs.google.com/document/d/1gEHGeOlVmWPC9necznYOUCMhgYxqykJKT_nZUMFJvWU/edit?disco=AAAAVD7naXo

See design document:
https://docs.google.com/document/d/1Exzy7oRUgMGdmwTGg9o1SQ9Lc1az-PbLnXCMpZwpFPw/edit?usp=sharing

Change-Id: I84ecbb10e4c3b52181f6efa07f1705e366f61ac0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3668390
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025697}