Postpone the notification callback creation to Unlock
This CL is postponing the reception of the process singleton
notification callback.

The ChromeProcessSingleton does not process callback notifications
until the call to Unlock. The current CL is proposing to pass
the callback to Unlock(...) to postpone as late as possible
the creation of the callback.

This change should almost be a no-op. The purpose of changing the
API is to allow the followup CLs to lift much more earlier the
creation of the process singleton.

Bug: 1249870
Change-Id: I8fddde1d8b31ce3e845e11e8af83d4e17f84bccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3764565
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025737}