[linux][lacros] Converting the tab hover card bubble into a tooltip...
...part 1.

This is the first part of [1] that was suggested by the revewer to be
split.

Historically Linux (and then Lacros because it inherited from Linux)
did not use the accelerated widgets for the bubbles due to some
rendering issues which are now gone.  That is going to change now,
starting with the tab hover card bubble, see the linked crbug.

Here the views::Widget::InitParams struct gets a new field specific to
Linux and Lacros that allows overriding the default widget type.  The
code in DesktopWindowTreeHostPlatform and ChromeViewsDelegateLinux is
updated to recognise the new field.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3689500

Bug: 1280332
Change-Id: I70ef0e0faee7b3a12f1bc49edd05f08b6ef7ddb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3768532
Auto-Submit: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1025781}