Fix RenderText API fuzzer runnign on CrOs over Linux
This CL is fixing a regression detected by clusterfuzz.

After this change:
https://chromium-review.googlesource.com/c/chromium/src/+/3715211

There was a path that was not triggered before since it's
fonts-dependent. The issue seems to always happening with
the mock fonts.

The test discardable memory should be initialized. The issue
was that it was not on CrOs on Linux.

Bug: 1345111
Change-Id: Ie395932fe95714cdc22ffb097d99356665776fae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3773894
Reviewed-by: Ben Wagner <bungeman@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025782}