Adjust wording for IDS_COMPONENTS_EVT_STATUS_ALREADY_UP_TO_DATE
The existing wording `Component not updated` seems to convey
that something went wrong with updating and the component is out
of date, when in fact it "wasn't updated" because the local
version is already up to date.

This CL changes the wording to more clearly convey the actual
state.

Bug: 1344867
Change-Id: I244542e6912d62e417a347d4bb38bbce2f087446
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3764703
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Auto-Submit: Bryant Chandler <bryantchandler@chromium.org>
Commit-Queue: Bryant Chandler <bryantchandler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025788}