Fix undefined behavior in inline script streamer
The V8 scanner expects a valid pointer to increment even when the length
of the buffer is 0. There were already tests that catch this when run
with ubsan, but it looks like none of the bots run unit tests in that
configuration.

We may want to fix this in V8, but since I want to merge this to M104
this is a simpler fix for now.

Bug: 1338382
Change-Id: Ifcabb81b447679ad15a1b9f93ecfc4fce86d9772
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3773337
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025789}