Implement service integration for CouponPersistedTabData
This CL implements the from method of CouponPersistedTabData which
integrates with the jetpack endpoint via EndpointFetcher and obtains a
response that is converted into a JSONArray that is parsed to create a
Coupon object with the coupon name and promo code. Future CLs will
use the discount amounts from the endpoint response to compose a String
that will be displayed on the Tab Switcher UI (see crrev.com/c/3749963).

Change-Id: I80da8f316e2c542807b9636dc33bf807ab39a79b
Bug: 1335109
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3724175
Reviewed-by: David Maunder <davidjm@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Wamia Said <wamia@google.com>
Cr-Commit-Position: refs/heads/main@{#1025821}