[CacheStorage] 3/ Use distinct bucket path for Background Fetch
This CL is the third in a set of CLs to integrate Cache Storage with the
Buckets API. The buckets integration is nearly complete in a WIP CL [1],
but we are splitting up the changes into smaller CLs for easier review.

Cache Storage uses the disk cache for storing data to disk, and
the disk cache subsystem expects that only one disk cache backend
instance is used in conjunction with any given directory at any
given time. Since we have distinct Cache Storage objects (with
separate disk cache backend instances) for the Cache Storage
API and the Background Fetch API, we also need to ensure
we compute distinct bucket paths for these as well (as is done
today by the ConstructOriginPath logic).

Design doc:
https://docs.google.com/document/d/1VDm8M4vRyxCpvp_m3Ojs18W9g1WpqvSOsKf0A6hnAFM/edit?pli=1#heading=h.1kp8e5sw3qha

Implementation design doc:
https://docs.google.com/document/d/1B_tUfTJDeuqRyehvX6aM44YwvOkKWVNM5LNBu3_leqU/edit?usp=sharing

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3735215

Bug: 1218097, 1304786
Change-Id: I23f81ee035e62fc339322a75dbac64393886e14e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3770176
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Andrew Williams <awillia@google.com>
Cr-Commit-Position: refs/heads/main@{#1025828}